+++ app/Parser/BeaverBuilder.php
@@ -21,7 +21,7 @@ use stdClass;
  */
 class BeaverBuilder extends Parser_Base implements Parser {
 	/**
-	 * Internal name of the parser.
+	 * The internal name of this parser.
 	 *
 	 * @var string
 	 */