+++ app/Parser/Bricks.php
@@ -19,7 +19,7 @@ use easyLanguage\EasyLanguage\Post_Object;
  */
 class Bricks extends Parser_Base implements Parser {
 	/**
-	 * Internal name of the parser.
+	 * The internal name of this parser.
 	 *
 	 * @var string
 	 */