+++ app/Parser/Elementor.php
@@ -22,7 +22,7 @@ use easyLanguage\EasyLanguage\Post_Object;
  */
 class Elementor extends Parser_Base implements Parser {
 	/**
-	 * Internal name of the parser.
+	 * The internal name of this parser.
 	 *
 	 * @var string
 	 */